Move update impl to NpmBasedUrlUpdater
